주문 통합검색시 order_products 기준으로 최신순 정렬 되도록 구현

Ref: #24

+++ src/main/java/com/midnear/midnearshopping/domain/vo/notice/NoticeImagesVo.java
@@ -10,7 +10,7 @@ import java.util.Date;
 @AllArgsConstructor
 @Builder
 public class NoticeImagesVo {
-    private Integer noticeImageId;
+    private Long noticeImageId;
 
     private String imageUrl;
 
@@ -20,7 +20,7 @@ public class NoticeImagesVo {
 
     private Date imageCreationDate;
 
-    private int noticeId;
+    private Long noticeId;
     public static NoticeImagesVo toEntity(NoticeImagesDto noticeImagesDto) {
         return NoticeImagesVo.builder()
                 .noticeImageId(noticeImagesDto.getNoticeImageId())
